requirements.txt
absl-py==2.1.0
academictorrents==2.3.3
accelerate==0.34.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.2
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
anykeystore==0.2
appdirs==1.4.4
argcomplete==3.4.0
arxiv_latex_cleaner==1.0.8
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
av==13.0.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bencode.py==2.0.0
bitstring==3.1.5
black==24.1.0
bleach==6.2.0
boto3==1.34.127
botocore==1.34.127
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
cmake==3.28.3
colorama==0.4.6
-e git+https://github.com/hpcaitech/ColossalAI.git@adae123df3badfb15d044bd416f0cf29f250bc86#egg=colossalai
contexttimer==0.3.3
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.11
dacite==1.7.0
DataProperty==1.0.1
datasets==2.16.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.30.1
dill==0.3.7
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
draccus==0.9.1
einops==0.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
evaluate==0.4.3
exceptiongroup==1.2.1
executing==2.2.0
fabric==3.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.21.1
ffmpy==0.3.2
fiftyone==0.24.1
fiftyone-brain==0.16.1
fiftyone_db==1.1.3
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
future==0.17.1
gast==0.4.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
glob2==0.7
google==3.0.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio_client==0.16.1
gradio_imageslider==0.0.20
graphql-core==3.2.3
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hf_transfer==0.1.8
hjson==3.1.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
huggingface-hub==0.26.1
humanize==4.9.0
hupper==1.12.1
Hypercorn==0.17.3
hyperframe==6.0.1
identify==2.5.35
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.4.0
inflate64==1.0.0
iniconfig==2.0.0
invoke==2.2.0
ipython==8.12.3
isort==5.13.2
jax==0.4.30
jaxlib==0.4.30
jaxtyping==0.2.33
jedi==0.19.2
Jinja2==3.0.3
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaleido==0.2.1
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.3
lazy_loader==0.3
libclang==18.1.1
lightning==2.2.3
lightning-utilities==0.11.0
lit==18.1.1
loguru==0.7.2
lxml==5.3.0
mamba-ssm==2.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.2
ml-dtypes==0.4.1
mongoengine==0.24.2
motor==3.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
multivolumefile==0.2.3
mutagen==1.47.0
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.8.0
numexpr==2.10.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.46.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.10.0
orjson==3.10.6
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parameterized==0.9.0
paramiko==3.4.0
parquet==1.3.1
parso==0.8.4
PasteDeploy==3.1.0
pathspec==0.12.1
pathvalidate==3.2.1
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
plotly==5.22.0
pluggy==1.5.0
ply==3.11
portalocker==2.10.1
pprintpp==0.4.0
pre-commit==3.6.2
priority==2.0.0
prompt_toolkit==3.0.50
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==0.21.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyav==12.0.5
pybcj==1.0.2
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
pymongo==4.7.3
PyNaCl==1.5.0
pynvml==11.5.0
pyparsing==3.1.2
pyppmd==1.1.0
PyPubSub==3.3.0
pyramid==2.0.2
PySocks==1.7.1
pytablewriter==1.2.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.2.0
pytorch-fid==0.3.0
pytorch-lightning==1.5.9
pytorch-wavelets==1.3.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyyaml-include==1.4.1
pyzmq==26.3.0
pyzstd==0.16.0
rarfile==4.2
ray==2.9.3
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.5.1
s3transfer==0.10.1
sacrebleu==2.4.3
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
sqlitedict==2.1.0
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.37.2
strawberry-graphql==0.138.1
svg.py==1.4.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
taskgroup==0.0.0a4
tcolorpy==0.1.6
tenacity==8.3.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
test-tube==0.7.5
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
thriftpy2==0.5.2
tifffile==2024.2.12
tiktoken==0.7.0
timm==0.9.10
tinycss2==1.4.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torch-fidelity==0.3.0
torch-summary==1.4.5
torchaudio==2.2.1
torchinfo==1.8.0
torchmetrics==1.3.2
torchvision==0.17.1
tornado==6.4.2
tqdm==4.66.2
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transaction==4.0
transformers==4.46.1
transformers-stream-generator==0.0.5
translationstring==1.4
triton==2.2.0
typeguard==2.13.3
typepy==1.3.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
universal-analytics-python3==1.1.1
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
velruse==1.1.1
venusian==3.1.0
virtualenv==20.26.2
voxel51-eta==0.12.6
wandb==0.17.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.86
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xmltodict==0.13.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
yt-dlp==2024.8.6
zipp==3.18.1
zope.deprecation==5.0
zope.interface==6.4.post2
zope.sqlalchemy==3.1
zss==1.2.0
zstandard==0.23.0
